+++ engine/src/main/java/nl/inl/blacklab/search/results/ResultsStats.java
@@ -7,27 +7,27 @@ public abstract class ResultsStats implements SearchResult {
 
     public static final ResultsStats SEARCH_NOT_STARTED_YET = new ResultsStats() {
         @Override
-        public boolean processedAtLeast(int lowerBound) {
+        public boolean processedAtLeast(long lowerBound) {
             return false;
         }
 
         @Override
-        public int processedTotal() {
+        public long processedTotal() {
             throw new RuntimeException("cannot access total, search not started");
         }
 
         @Override
-        public int processedSoFar() {
+        public long processedSoFar() {
             return 0;
         }
 
         @Override
-        public int countedSoFar() {
+        public long countedSoFar() {
             return 0;
         }
 
         @Override
-        public int countedTotal() {
+        public long countedTotal() {
             throw new RuntimeException("cannot access total, search not started");
         }
 
@@ -47,24 +47,24 @@ public abstract class ResultsStats implements SearchResult {
         }
     };
 
-    public abstract boolean processedAtLeast(int lowerBound);
+    public abstract boolean processedAtLeast(long lowerBound);
 
     /**
      * This is an alias of resultsProcessedTotal().
      *
      * @return number of hits processed total
      */
-    public int size() {
+    public long size() {
         return processedTotal();
     }
 
-    public abstract int processedTotal();
+    public abstract long processedTotal();
 
-    public abstract int processedSoFar();
+    public abstract long processedSoFar();
 
-    public abstract int countedSoFar();
+    public abstract long countedSoFar();
 
-    public abstract int countedTotal();
+    public abstract long countedTotal();
 
     public abstract boolean done();
 
@@ -118,7 +118,7 @@ public abstract class ResultsStats implements SearchResult {
      * @return
      */
     @Override
-    public int numberOfResultObjects() {
+    public long numberOfResultObjects() {
         return 1;
     }
